refactor: entities에서 entities/location/으로 폴더 위치 변경
