+++ frontend/src/entities/location/api/types/RecommendationResultAPITypes.ts
