+++ frontend/src/entities/location/contexts/LocationsContext.tsx
@@ -1,6 +1,6 @@
 import { createContext } from 'react';
 
-import { useLocationsReturn } from '@entities/hooks/useLocations';
+import { useLocationsReturn } from '@entities/location/hooks/useLocations';
 
 const LocationsContext = createContext<useLocationsReturn | null>(null);
