+++ frontend/src/entities/location/types/Location.ts
