+++ frontend/src/entities/location/types/LocationRequirement.ts
