+++ frontend/src/features/recommendation/components/spotItemList/SpotItemList.tsx
@@ -1,4 +1,4 @@
-import { RecommendedLocation } from '@entities/types/Location';
+import { RecommendedLocation } from '@entities/location/types/Location';
 
 import SpotItem from '@shared/components/spotItem/SpotItem';
 import { flex } from '@shared/styles/default.styled';
