+++ frontend/src/features/recommendation/types/SelectedLocation.ts
@@ -1,3 +1,3 @@
-import { RecommendedLocation } from '@entities/types/Location';
+import { RecommendedLocation } from '@entities/location/types/Location';
 
 export type SelectedLocation = RecommendedLocation | null;
