+++ frontend/src/mocks/LocationsRequestBodyMock.ts
@@ -1,4 +1,4 @@
-import { RecommendationRequestBody } from '@entities/api/types/RecommendationIdAPITypes';
+import { RecommendationRequestBody } from '@entities/location/api/types/RecommendationIdAPITypes';
 
 export const LocationsRequestBodyMock: RecommendationRequestBody = {
   startingPlaceNames: ['강변역', '동대문역', '서울대입구역'],
