+++ frontend/src/shared/components/startingSpotWrapper/StartingSpotWrapper.tsx
@@ -1,4 +1,4 @@
-import { StartingPlace } from '@entities/types/Location';
+import { StartingPlace } from '@entities/location/types/Location';
 
 import StartingSpotName from '@shared/components/startingSpotName/StartingSpotName';
 import { flex, typography } from '@shared/styles/default.styled';
